правки от Кирилла

+++ erSMS_v2.x/Form1.vb
@@ -12,15 +12,13 @@ Public Class Form1
 
 #Region "DB presets"
     Dim connectionString As String = "Data Source={0};Version=3;"
-    Dim DB As String = IO.Directory.GetCurrentDirectory & "\Plugins\erSMS_Resources\settings.db"
-    'Dim SQL As String
+    Dim DB As String = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) & "\CLM\Plugins\erSMS_Resources\settings.db"
     Dim m_dbConn As SQLiteConnection
     Dim m_sqlCmd As SQLiteCommand
 #End Region
 
     Private Sub Form1_Load(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles MyBase.Load
         connectionString = String.Format(connectionString, DB)
-        Connect()
         If args(0) = "Setting" Then
             Me.Show()
             AddCOM()
@@ -29,6 +27,7 @@ Public Class Form1
                 Try
                     'создать БД с настройками по умолчанию
                     SQLiteConnection.CreateFile(DB)
+                    Connect()
                     DBSend("CREATE TABLE IF NOT EXISTS phones (number TEXT, gate TEXT)")
                     DBSend("CREATE TABLE IF NOT EXISTS modems (com TEXT, name TEXT)")
                     DBSend("CREATE TABLE IF NOT EXISTS killthemall (killme TEXT)")
@@ -36,8 +35,9 @@ Public Class Form1
                     For i = 0 To killme.Count - 1
                         DBSend("INSERT INTO killthemall (killme) VALUES ('" & killme(i) & "')")
                     Next
+                    LogMrg("БД успешно создана!")
                 Catch ex As Exception
-                    LogMrg("Creating default DB error: " & ex.Message)
+                    LogMrg("Creating_default_DB_error: " & ex.Message)
                 End Try
             ElseIf My.Computer.FileSystem.FileExists(DB) = True Then
                 'просто считать настройки с существующей БД
@@ -49,6 +49,7 @@ Public Class Form1
             Timer1.Start()
         ElseIf args.Count >= 1 And args(0) <> "Setting" Then
             Me.Hide()
+            Connect()
             BornToKill()
             Dim message As String = Nothing
             For i = 0 To args.Count - 1
@@ -66,7 +67,7 @@ Public Class Form1
                     SMS(number.Rows(i).Item(0), gate.Rows(0).Item(0), message)
                 Next
             Catch ex As Exception
-                LogMrg("Sending error: " & ex.Message)
+                LogMrg("Sending_error: " & ex.Message)
             End Try
             Me.Close()
         ElseIf args.Count = 0 Then
@@ -84,7 +85,7 @@ Public Class Form1
             m_dbConn = New SQLiteConnection(connectionString)
             m_dbConn.Open()
         Catch ex As Exception
-            LogMrg("Error: " + ex.Message)
+            LogMrg("Connect_to_DB_Error: " + ex.Message)
         End Try
     End Sub
 
@@ -96,7 +97,7 @@ Public Class Form1
             m_sqlCmd.CommandText = sqlQuery
             m_sqlCmd.ExecuteNonQuery()
         Catch ex As Exception
-            LogMrg("BDSend Error: " + ex.Message)
+            LogMrg("BDSend_Error: " + ex.Message)
         End Try
     End Sub
 
@@ -114,7 +115,7 @@ Public Class Form1
                 Next
             End If
         Catch ex As Exception
-            LogMrg("DBPhoneRead Error: " + ex.Message)
+            LogMrg("DBPhoneRead_Error: " + ex.Message)
         End Try
     End Sub
 
@@ -130,7 +131,7 @@ Public Class Form1
                 ComboBox3.Items.Add(modem.Rows(i).Item(1))
             Next
         Catch ex As Exception
-            LogMrg("ModemRead error: " & ex.Message)
+            LogMrg("ModemRead_error: " & ex.Message)
         End Try
     End Sub
 
@@ -189,7 +190,7 @@ Public Class Form1
             'логгирование
             LogMrg(sp_result)
         Catch ex As Exception
-            LogMrg("SMS error: " & ex.Message)
+            LogMrg("SMS_error: " & ex.Message)
         End Try
     End Sub
 
@@ -221,7 +222,7 @@ Public Class Form1
         'End Try
         Try
             Dim D As Date = Now
-            My.Computer.FileSystem.WriteAllText(IO.Directory.GetCurrentDirectory & "\Plugins\erSMS_Resources\ersms.log", D & ": " & log.Replace(New Char() {vbLf, vbCr}, "").TrimEnd(New Char() {vbLf, vbCr}) & vbNewLine, True)
+            My.Computer.FileSystem.WriteAllText(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) & "\CLM\Plugins\erSMS_Resources\ersms.log", D & ": " & log.Replace(New Char() {vbLf, vbCr}, "").TrimEnd(New Char() {vbLf, vbCr}) & vbNewLine, True)
         Catch ex As Exception
         End Try
     End Sub
@@ -321,8 +322,17 @@ Public Class Form1
 
     'очистка базы тел-модем
     Private Sub Button3_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles Button3.Click
-        DBSend("DELETE FROM phones")
-        DBPhoneRead()
+        Dim response As MsgBoxResult
+        Dim style As MsgBoxStyle = MsgBoxStyle.DefaultButton2 Or _
+           MsgBoxStyle.Critical Or MsgBoxStyle.YesNo
+        ' Display message.
+        response = MsgBox("Будет очищен список связок 'телефон-шлюз', продолжить?", style, "Очистка списка шлюзов")
+        If response = MsgBoxResult.Yes Then   ' User chose Yes.
+            DBSend("DELETE FROM phones")
+            DBPhoneRead()
+        Else
+            ' Perform some other action.
+        End If
     End Sub
 
     'мониторинг модемов
@@ -335,7 +345,7 @@ Public Class Form1
     Private Sub GateListUpdate()
         AddCOM()
         Dim modem As DataTable = New DataTable()
-        'Connect()
+        Connect()
         ModemRead()
         Dim OpSos As String = ""
         Dim SignalInfo As String = ""
@@ -356,7 +366,7 @@ Public Class Form1
                 name = modem.Rows(i).Item(1)
                 com = modem.Rows(i).Item(0)
                 Dim sigarr() As String = SignalInfo.Split(" ")
-                DataGridView2.Rows.Add(name, com, OpSos, sigarr(0))
+                DataGridView2.Rows.Add(name, com, OpSos, sigarr(0) & " dBm")
                 DataGridView2.Rows(i).Cells(4).Value = New Bitmap(IO.Directory.GetCurrentDirectory & "\Plugins\erSMS_Resources\Signal\" & sigarr(1))
                 ComboBox2.Items.Add(name)
             Next
@@ -474,6 +484,7 @@ Public Class Form1
         ' Display message.
         response = MsgBox("Будет очищена база шлюзов, продолжить?", style, "Очистка списка шлюзов")
         If response = MsgBoxResult.Yes Then   ' User chose Yes.
+            Connect()
             DBSend("DELETE FROM modems")
             GateListUpdate()
         Else
@@ -493,6 +504,7 @@ Public Class Form1
             MsgBox("Введите название процесса подлежащего завершению!")
             Exit Sub
         Else
+            Connect()
             DBSend("INSERT INTO killthemall (killme) VALUES ('" & TextBox3.Text & "')")
             TextBox3.Clear()
             KillListboxUpdate()
@@ -501,8 +513,18 @@ Public Class Form1
 
     'очистить список смертников
     Private Sub Button9_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles Button9.Click
-        DBSend("DELETE FROM killthemall")
-        KillListboxUpdate()
+        Dim response As MsgBoxResult
+        Dim style As MsgBoxStyle = MsgBoxStyle.DefaultButton2 Or _
+           MsgBoxStyle.Critical Or MsgBoxStyle.YesNo
+        ' Display message.
+        response = MsgBox("Будет очищен список процессов подлежащих уничтожению, продолжить?", style, "Очистка списка шлюзов")
+        If response = MsgBoxResult.Yes Then   ' User chose Yes.
+            Connect()
+            DBSend("DELETE FROM killthemall")
+            KillListboxUpdate()
+        Else
+            ' Perform some other action.
+        End If
     End Sub
 
     'удалить одного приговорённого
@@ -513,6 +535,7 @@ Public Class Form1
 
     'отправка АТ-команды
     Private Sub Button10_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles Button10.Click
+        Connect()
         Dim at As String = ComboBox2.SelectedItem
         Try
             Dim send As DataTable = New DataTable()
@@ -533,7 +556,7 @@ Public Class Form1
     'показать лог
     Private Sub ПоказатьЛогToolStripMenuItem_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles ПоказатьЛогToolStripMenuItem.Click
         Try
-            Process.Start(IO.Directory.GetCurrentDirectory & "\Plugins\erSMS_Resources\ersms.log")
+            Process.Start(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) & "\CLM\Plugins\erSMS_Resources\ersms.log")
         Catch ex As Exception
             MsgBox(ex.Message)
         End Try
@@ -542,5 +565,35 @@ Public Class Form1
     'выход
     Private Sub ВыходToolStripMenuItem_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles ВыходToolStripMenuItem.Click
         End
+    End Sub
+
+    'очистить терминал
+    Private Sub Button11_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles Button11.Click
+        ListBox2.Items.Clear()
+    End Sub
+
+    'показать ман
+    Private Sub ИнструкцияToolStripMenuItem_Click(ByVal sender As System.Object, ByVal e As System.EventArgs)
+
+    End Sub
+
+    'о программе
+    Private Sub ОПрограммеToolStripMenuItem_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles ОПрограммеToolStripMenuItem.Click
+        MsgBox("erSMS v2.0 - встраиваемое программное обеспечение" & vbNewLine & _
+               "для отправки SMS сообщений с помощью GSM устройств." & vbNewLine & _
+               "Ревчук Ю.Ю. 2018г." & vbNewLine & _
+               " https://vk.com/r2y_25rus" & vbNewLine & _
+               "Разработано для ПО Chanell Live Monitor:" & vbNewLine & _
+               "https://vk.com/clmonitor" & vbNewLine & _
+               "http://clmonitor.ru/")
+    End Sub
+
+    'показать лицензию
+    Private Sub ЛицензияToolStripMenuItem_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles ЛицензияToolStripMenuItem.Click
+        Try
+            Process.Start(IO.Directory.GetCurrentDirectory & "\Plugins\erSMS_Resources\LICENSE.txt")
+        Catch ex As Exception
+            MsgBox(ex.Message)
+        End Try
     End Sub
 End Class
